web-sockets-server/requirements.txt
kafka-python==1.4.4
websockets==7.0
asgiref==2.3.2
